* chore: move Avatar Group into Avatar package and remove unneeded JSON
