--- draft-packages/avatar-group/docs/companyAvatarData.json
@@ -1,17 +1,0 @@
-[
-  {
-    "fullName": "Hooli",
-    "avatarSrc": "https://d1e7r7b0lb8p4d.cloudfront.net/third-party-logos/hooli-logo.svg",
-    "isCompany": true
-  },
-  {
-    "fullName": "Looli",
-    "avatarSrc": "https://d1e7r7b0lb8p4d.cloudfront.net/third-party-logos/hooli-logo.svg",
-    "isCompany": true
-  },
-  {
-    "fullName": "Dooli",
-    "avatarSrc": "https://d1e7r7b0lb8p4d.cloudfront.net/third-party-logos/hooli-logo.svg",
-    "isCompany": true
-  }
-]
