--- draft-packages/avatar-group/docs/defaultAvatarData.json
@@ -1,38 +1,0 @@
-[
-  {
-    "fullName": "Adirana Aniseed",
-    "disableInitials": false,
-    "avatarSrc": "https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png",
-    "isCurrentUser": false
-  },
-  {
-    "fullName": "Bethany Blueberry",
-    "disableInitials": false,
-    "avatarSrc": "",
-    "isCurrentUser": false
-  },
-  {
-    "fullName": "Carey Cringle",
-    "disableInitials": false,
-    "avatarSrc": "https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png",
-    "isCurrentUser": false
-  },
-  {
-    "fullName": "Derrick Doolittle",
-    "disableInitials": false,
-    "avatarSrc": "",
-    "isCurrentUser": false
-  },
-  {
-    "fullName": "Evan Eavesdrop",
-    "disableInitials": false,
-    "avatarSrc": "",
-    "isCurrentUser": false
-  },
-  {
-    "fullName": "Fern Furlow",
-    "disableInitials": false,
-    "avatarSrc": "https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png",
-    "isCurrentUser": false
-  }
-]
