--- draft-packages/avatar-group/index.ts
@@ -1,1 +1,0 @@
-export * from "./KaizenDraft/AvatarGroup/AvatarGroup"
