--- draft-packages/avatar-group/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@kaizen/draft-avatar-group",
-  "version": "1.0.0",
-  "description": "The draft avatar group component",
-  "scripts": {
-    "prepublish": "tsc --project tsconfig.dist.json",
-    "build": "yarn clean && yarn prepublish",
-    "build:watch": "yarn clean && yarn prepublish --watch",
-    "clean": "rimraf '**/*.d.ts' '**/*.js' '**/*.map'"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/cultureamp/kaizen-design-system.git",
-    "directory": "draft-packages/avatar-group"
-  },
-  "bugs": {
-    "url": "https://github.com/cultureamp/kaizen-design-system/issues"
-  },
-  "files": [
-    "**/*",
-    "!**/*.ts",
-    "!**/*.tsx",
-    "**/*.d.ts",
-    "!**/*.spec.*",
-    "!**/*.snap",
-    "!draft-templates",
-    "!stories",
-    "!docs",
-    "!tsconfig.dist.json"
-  ],
-  "author": "",
-  "private": false,
-  "license": "MIT",
-  "dependencies": {
-    "@kaizen/component-library": "^11.5.6",
-    "@kaizen/draft-avatar": "^2.0.0",
-    "@types/classnames": "^2.3.0",
-    "classnames": "^2.3.1",
-    "uuid": "^3.4.0"
-  },
-  "peerDependencies": {
-    "@kaizen/design-tokens": "^2.10.3 || ^3.0.0",
-    "react": "^16.9.0"
-  }
-}
