--- draft-packages/avatar-group/tsconfig.dist.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.dist.json",
-  "compilerOptions": {
-    "allowJs": false,
-    "declaration": true,
-    "noEmit": false,
-    "sourceMap": true
-  },
-  "exclude": ["docs"]
-}
