+++ draft-packages/avatar/docs/avatarGroupData.json
@@ -1,0 +1,57 @@
+{
+  "users": [
+    {
+      "fullName": "Adirana Aniseed",
+      "disableInitials": false,
+      "avatarSrc": "https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png",
+      "isCurrentUser": false
+    },
+    {
+      "fullName": "Bethany Blueberry",
+      "disableInitials": false,
+      "avatarSrc": "",
+      "isCurrentUser": false
+    },
+    {
+      "fullName": "Carey Cringle",
+      "disableInitials": false,
+      "avatarSrc": "https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png",
+      "isCurrentUser": false
+    },
+    {
+      "fullName": "Derrick Doolittle",
+      "disableInitials": false,
+      "avatarSrc": "",
+      "isCurrentUser": false
+    },
+    {
+      "fullName": "Evan Eavesdrop",
+      "disableInitials": false,
+      "avatarSrc": "",
+      "isCurrentUser": false
+    },
+    {
+      "fullName": "Fern Furlow",
+      "disableInitials": false,
+      "avatarSrc": "https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png",
+      "isCurrentUser": false
+    }
+  ],
+  "company": [
+    {
+      "fullName": "Hooli",
+      "avatarSrc": "https://d1e7r7b0lb8p4d.cloudfront.net/third-party-logos/hooli-logo.svg",
+      "isCompany": true
+    },
+    {
+      "fullName": "Looli",
+      "avatarSrc": "https://d1e7r7b0lb8p4d.cloudfront.net/third-party-logos/hooli-logo.svg",
+      "isCompany": true
+    },
+    {
+      "fullName": "Dooli",
+      "avatarSrc": "https://d1e7r7b0lb8p4d.cloudfront.net/third-party-logos/hooli-logo.svg",
+      "isCompany": true
+    }
+  ]
+}
