+++ draft-packages/avatar/package.json
@@ -35,7 +35,8 @@
     "@kaizen/component-library": "^11.5.8",
     "@types/classnames": "^2.3.0",
     "classnames": "^2.3.1",
-    "react-textfit": "^1.1.1"
+    "react-textfit": "^1.1.1",
+    "uuid": "^8.3.2"
   },
   "peerDependencies": {
     "@kaizen/design-tokens": "^2.10.3 || ^3.0.0",
